requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Message Queue & Caching
redis==5.0.1
kafka-python==2.0.2

# Data Processing
apache-beam==2.51.0
pyspark==3.5.0

# GraphQL
strawberry-graphql[fastapi]==0.215.1

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1